Reorganize conversions
This reorganizes the conversions as follows:
* Moves the headers and implementations from
  `emitc/Dialect/EmitC/Conversion` to `emitc/Conversion`.
* Creates individual headers for each conversion

There is more clean up to do which will be done in follow ups.
