+++ include/emitc/CMakeLists.txt
@@ -1,1 +1,2 @@
+add_subdirectory(Conversion)
 add_subdirectory(Dialect)
