+++ include/emitc/Dialect/EmitC/CMakeLists.txt
@@ -1,2 +1,1 @@
-add_subdirectory(Conversion)
 add_subdirectory(Transforms)
