+++ include/emitc/InitPasses.h
@@ -14,7 +14,7 @@
 #ifndef EMITC_INITPASSES_H
 #define EMITC_INITPASSES_H
 
-#include "emitc/Dialect/EmitC/Conversion/Passes.h"
+#include "emitc/Conversion/Passes.h"
 #include "emitc/Dialect/EmitC/Pipelines.h"
 #include "emitc/Dialect/EmitC/Transforms/Passes.h"
 #include "mlir/Conversion/Passes.h"
