+++ lib/Conversion/PassDetail.h
@@ -18,7 +18,7 @@ namespace emitc {
 class EmitCDialect;
 
 #define GEN_PASS_CLASSES
-#include "emitc/Dialect/EmitC/Conversion/Passes.h.inc"
+#include "emitc/Conversion/Passes.h.inc"
 
 } // namespace emitc
 } // namespace mlir
